Factor out the cpython._files module.
